requirements.txt
absl_py==2.1.0
akro==0.0.8
apex==0.9.10dev
atomics==1.0.3
av==10.0.0
Box2D==2.3.10
cffi==1.15.1
click==8.0.4
cloudpickle==3.1.1
cuml==0.6.1.post1
Cython==0.29.32
diffusers==0.11.1
dill==0.3.5.1
dm_control==1.0.9
dm_env==1.6
easydict==1.13
einops==0.4.1
fasteners==0.15
filelock==3.17.0
Flask==3.1.0
fvcore==0.1.5.post20221221
garage==2021.3.0
glfw==1.12.0
gymnasium==1.0.0
h5py==3.13.0
imagecodecs==2022.9.26
imagehash==4.3.2
imageio==2.22.0
importlib_metadata==8.0.0
importlib_metadata==8.6.1
ipdb==0.13.13
joblib==1.4.2
lpips==0.1.4
lz4==4.4.3
matplotlib==3.6.1
ml_collections==1.0.0
mmcv==2.2.0
mmdet==3.3.0
mmseg==1.3.0
numba==0.53.1
numcodecs==0.10.2
numpy==1.24.0
omegaconf==2.2.3
open3d==0.19.0
openvr==2.5.101
pandas==2.2.3
Pillow==9.2.0
Pillow==11.1.0
plotly==6.0.0
psutil==5.9.4
psutil==5.9.2
pycparser==2.22
pygame==2.1.2
pyglet==2.1.2
pymunk==6.2.1
pynput==1.7.7
PyOpenGL==3.1.9
PyOpenGL==3.1.9
pyrealsense2==2.55.1.6486
pytest==8.3.4
PyYAML==6.0.2
PyYAML==6.0.2
r3m==0.0.0
ray==2.2.0
sam2==1.1.0
sapien==2.2.2
scikit_video==1.1.11
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
spnav==0.9
submitit==1.5.2
tabulate==0.9.0
tensorflow==2.17.1
tensorflow_gan==2.1.0
termcolor==2.5.0
tf_agents==0.19.0
threadpoolctl==3.5.0
torch==1.12.1
torchmetrics==1.6.1
torchvision==0.13.1
tqdm==4.64.1
trajopt==0.0.1
transforms3d==0.4.2
xformers==0.0.29.post3
zarr==2.12.0
